+++ DICOM/IO/StreamByteSource.cs
@@ -41,13 +41,14 @@ namespace Dicom.IO
         /// Initializes a new instance of <see cref="StreamByteSource"/>.
         /// </summary>
         /// <param name="stream">Stream to read from.</param>
-        public StreamByteSource(Stream stream, FileReadOption readOption = FileReadOption.LargeOnDemand)
+        public StreamByteSource(Stream stream, FileReadOption readOption = FileReadOption.Default)
         {
             _stream = stream;
             _endian = Endian.LocalMachine;
             _reader = EndianBinaryReader.Create(_stream, _endian);
             _mark = 0;
-            _readOption = readOption;
+            // here the mapping of the default option is applied - may be extracted into some GlobalSettings class or similar
+            _readOption = (readOption == FileReadOption.Default) ? FileReadOption.ReadLargeOnDemand : readOption;
 
             LargeObjectSize = 64 * 1024;
 
@@ -168,14 +169,14 @@ namespace Dicom.IO
             {
                 buffer = EmptyBuffer.Value;
             }
-            else if (count >= LargeObjectSize && _readOption == FileReadOption.LargeOnDemand)
+            else if (count >= LargeObjectSize && _readOption == FileReadOption.ReadLargeOnDemand)
             {
                 buffer = new StreamByteBuffer(_stream, _stream.Position, count);
                 _stream.Seek((int)count, SeekOrigin.Current);
             }
             else if (count >= LargeObjectSize && _readOption == FileReadOption.SkipLargeTags)
             {
-                buffer = EmptyBuffer.Value;
+                buffer = null;
             }
             else // count < LargeOpjectSize || _readOption == FileReadOption.ReadAll
             {